requirements.txt
python-dotenv==1.0.0
Telethon==1.26.1
wolframalpha==5.0
aiohttp==3.11.13
pandas==2.2.3
pyarrow==19.0.1
faiss-cpu==1.10.0
g4f==0.4.7.7